+++ CHANGELOG.md
@@ -90,6 +90,7 @@
 - Port the logout and relogin controllers to Symfony !3682 @alex.simm
 - Refactored join info page to vue !3737 @chriswalg
 - Clean up code for maintenance and background tasks !3735 @\_fridtjof_
+- Port the dashboard controller to Symfony !3736 @\_fridtjof_
 
 ## Dev/Test/CI stuff
 - Improved API tests runtime !3506 @alex.simm